+++ src/Control/Distributed/Spark/RDD.hs
@@ -38,6 +38,7 @@ module Control.Distributed.Spark.RDD
   , union
   , sortBy
   , sample
+  , randomSplit
   , first
   , getNumPartitions
   , saveAsTextFile
@@ -48,11 +49,14 @@ module Control.Distributed.Spark.RDD
 import Prelude hiding (filter, map, subtract, take)
 import Control.Distributed.Closure
 import Control.Distributed.Spark.Closure (reflectFun)
+import Control.Monad
 import Data.Choice (Choice)
 import qualified Data.Choice as Choice
 import Data.Int
 import qualified Data.Text as Text
 import Data.Typeable (Typeable)
+import Data.Vector.Storable as V (fromList)
+import Foreign.JNI
 import Language.Java
 import Language.Java.Inline
 -- We don't need this instance. But import to bring it in scope transitively for users.
@@ -216,6 +220,16 @@ sample
   -> Double -- ^ fraction of elements to keep
   -> IO (RDD a)
 sample rdd replacement frac = [java| $rdd.sample($replacement, $frac) |]
+
+randomSplit
+  :: RDD a
+  -> [Double] -- ^ Statistical weights of RDD fractions.
+  -> IO [RDD a]
+randomSplit rdd weights = do
+  jweights <- reflect $ V.fromList weights
+  arr :: JObjectArray <- [java| $rdd.randomSplit($jweights) |]
+  n <- getArrayLength arr
+  forM [0 .. n - 1] (getObjectArrayElement arr)
 
 first :: Reify a => RDD a -> IO a
 first rdd = do