+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
 
+* More `RDD` method bindings: `randomSplit`.
+
 ## [0.7.2] - 2017-12-25
 
 ### Added